+++ Source/Common/SQLLiteReport.h
@@ -73,11 +73,11 @@ public:
     virtual int  save_report(int user, long file_id, MediaConchLib::report reportKind, MediaConchLib::format format,
                              const std::string& options,
                              const std::string& report, MediaConchLib::compression,
-                             int mil_version, std::string& err);
+                             unsigned long long int mil_version, std::string& err);
     virtual int  update_report(int user, long file_id, MediaConchLib::report reportKind, MediaConchLib::format format,
                                const std::string& options,
                                const std::string& report, MediaConchLib::compression,
-                               int mil_version, std::string& err);
+                               unsigned long long int mil_version, std::string& err);
     virtual int  remove_report(int user, long filename, std::string& err);
     virtual int  remove_all_reports(int user, std::string& err);
     virtual int  get_report(int user, long file_id, MediaConchLib::report reportKind, MediaConchLib::format format,
@@ -86,7 +86,7 @@ public:
     virtual int  report_is_registered(int user, long file_id, MediaConchLib::report reportKind,
                                       MediaConchLib::format format, const std::string& options,
                                       bool& registered, std::string& err);
-    virtual int  version_registered(int user, long file_id, std::string& err);
+    virtual unsigned long long int version_registered(int user, long file_id, std::string& err);
     virtual int  get_elements(int user, std::vector<std::string>& vec, std::string& err);
     virtual int  get_elements(int user, std::vector<long>& vec, std::string& err);
     virtual int  get_element_report_kind(int user, long file_id, MediaConchLib::report& report_kind, std::string& err);
